fixed doctor-general, filtering on empty makes for bad docs
