+++ packages/doctor-general/comment.js
@@ -1,15 +1,17 @@
-import { extname, basename, dirname, join as pathJoin } from 'node:path'
-import { capitalToKebab, stripLeadingHyphen } from './string'
-import { isNotEmpty } from 'inherent'
-import { TESTABLE_EXAMPLE } from './constants'
+import { join as pathJoin } from 'node:path'
+import { writeFileWithAutoPath } from 'file-system'
+import { getImportsForTests, filterAndStructureTests } from './comment-test'
+import { filterAndStructureComments } from './comment-documentation'
+import { commentToMarkdown } from './renderer-markdown'
+import { commentToJestTest } from './renderer-jest'
+import { prepareMetaFiles } from './next-meta-files'
+import { log } from './log'
+import { parallel } from 'fluture'
 import {
-  includes,
-  prop,
   __ as $,
   addIndex,
   always as K,
   any,
-  anyPass,
   applySpec,
   chain,
   cond,
@@ -19,17 +21,18 @@ import {
   equals,
   filter,
   findIndex,
+  flatten,
   fromPairs,
+  groupBy,
   head,
   identity as I,
   ifElse,
   isEmpty,
+  join,
   last,
   length,
   map,
   match,
-  path,
-  pathOr,
   pipe,
   propOr,
   reduce,
@@ -38,20 +41,15 @@ import {
   slice,
   split,
   startsWith,
+  tap,
   test as testRegExp,
-  toLower,
   toPairs,
   trim,
   uniq,
 } from 'ramda'
-import { findJSDocKeywords, cleanupKeywords } from './file'
-import {
-  unlines,
-  formatComment,
-  trimComment,
-  wipeComment,
-  stripRelative,
-} from './text'
+import { cleanFilename, findJSDocKeywords, cleanupKeywords } from './file'
+import { formatComment, trimComment, wipeComment } from './text'
+import { unlines } from 'knot'
 
 const linkRegex = /\{@link (.*)+\}/g
 export const matchLinks = pipe(
@@ -195,13 +193,6 @@ const summarize = lines => {
   )(stripped)
 }
 
-export const slug = name => {
-  const slashPlus = name.lastIndexOf('/') + 1
-  return name.indexOf('.') > -1
-    ? name.slice(slashPlus, name.indexOf('.'))
-    : name.slice(slashPlus)
-}
-
 // getFileGroup :: String -> Comment
 const getFileGroup = propOr('', 'group')
 const addTo = propOr('', 'addTo')
@@ -254,139 +245,80 @@ export const getExample = curry((file, end, i) =>
   )(file)
 )
 
-// isJSDocComment :: String -> Boolean
-export const isJSDocComment = pipe(
-  trim,
-  anyPass([startsWith('/**'), startsWith('*'), startsWith('*/')])
+const renderFile = curry((testMode, file) => {
+  const importsForTests = getImportsForTests(testMode, file)
+  return pipe(
+    map(testMode ? commentToJestTest : commentToMarkdown),
+    z => {
+      const out = !testMode
+        ? ['# ' + file.slugName, file.pageSummary, ...z]
+        : [
+            '// This test automatically generated by doctor-general.',
+            `// Sourced from '${file.filename}', edits to this file may be erased.`,
+            `import {
+  ${importsForTests.join(',\n  ')}
+} from '../${file.slugName}'\n`,
+            ...z,
+          ]
+      return out
+    },
+    join(testMode ? '\n' : '\n\n')
+  )(file.comments)
+})
+
+export const writeCommentsToFiles = curry(({ testMode, outputDir }, x) =>
+  pipe(
+    toPairs,
+    map(([workspace, commentedFiles]) => {
+      const filesToWrite = map(file => {
+        const filePathToWrite = pathJoin(
+          outputDir,
+          workspace,
+          // this part is the structure of the file we wanna write
+          cleanFilename(testMode, file)
+        )
+
+        return writeFileWithAutoPath(
+          filePathToWrite,
+          renderFile(testMode, file)
+        )
+      })(commentedFiles)
+      return testMode
+        ? filesToWrite
+        : filesToWrite.concat(
+            prepareMetaFiles(testMode, outputDir, workspace, commentedFiles)
+          )
+    }),
+    flatten,
+    parallel(10)
+  )(x)
 )
 
-export const combineFiles = curry((leftToRight, a, b) =>
-  !leftToRight
-    ? combineFiles(true, b, a)
-    : {
-        ...a,
-        ...b,
-        comments: [...a.comments, ...b.comments],
-        links: [...a.links, ...b.links],
-      }
+export const renderComments = curry((testMode, outputDir, x) =>
+  chain(
+    pipe(
+      groupBy(propOr('unknown', testMode ? 'testPath' : 'workspace')),
+      tap(xxx =>
+        log.core(
+          `grouped ${testMode ? 'tests' : 'comments'}...`,
+          pipe(
+            toPairs,
+            map(
+              ([k, y]) =>
+                `\n${k}:\n - ${pipe(
+                  map(z => z.slugName),
+                  join('\n - ')
+                )(y)}`
+            ),
+            join('\n')
+          )(xxx)
+        )
+      ),
+      writeCommentsToFiles({ testMode, outputDir })
+    )
+  )(x)
 )
 
-export const parsePackageName = y => {
-  const slash = y.indexOf('/')
-  const start = slash + 1
-  const end = y.indexOf('/', start)
-  return y.slice(start, end)
-}
-
-export const filterAndStructureComments = pipe(
-  filter(pipe(propOr([], 'comments'), isEmpty)),
-  map(raw => {
-    const filename = stripRelative(raw.filename)
-    return {
-      ...raw,
-      comments: raw.comments.map(r => ({ ...r, filename })),
-      filename,
-      workspace: parsePackageName(filename),
-    }
-  }),
-  reduce((agg, file) => {
-    const filenames = map(prop('filename'), agg)
-    const alreadyInList = filenames.includes(file.filename)
-    const anyFile = file.comments.filter(({ structure }) => structure.asFile)
-    const someFile = anyFile.length > 0 ? anyFile[0] : false
-    const asFilePath = pipe(
-      defaultTo({}),
-      pathOr('???', ['structure', 'asFile'])
-    )(someFile)
-    const withOrder = pipe(pathOr('0', ['structure', 'order']), x =>
-      parseInt(x, 10)
-    )(someFile)
-    const dir = dirname(file.filename)
-    const newFile = someFile ? pathJoin(dir, asFilePath) : '???'
-    return alreadyInList
-      ? map(raw => {
-          const check = raw.filename === file.filename
-          return check ? combineFiles(raw.order < withOrder, raw, file) : raw
-        })(agg)
-      : [
-          ...agg,
-          someFile
-            ? {
-                ...file,
-                filename: newFile,
-                order: withOrder,
-                originalFilename: file.filename,
-              }
-            : file,
-        ]
-  }, [])
-)
-
-export const hasTestableExample = pipe(
-  pathOr('', ['structure', 'example']),
-  includes(TESTABLE_EXAMPLE)
-)
-export const filterAndStructureTests = pipe(
-  filter(pipe(propOr([], 'comments'), filter(hasTestableExample), isNotEmpty)),
-  map(raw => {
-    const filename = stripRelative(raw.filename)
-    const ext = extname(filename)
-    return {
-      ...raw,
-      comments: raw.comments.map(r => ({ ...r, filename })),
-      filename,
-      testPath: `${basename(filename, ext)}.spec${ext}`,
-    }
-  }),
-  reduce((agg, file) => {
-    const filenames = map(prop('filename'), agg)
-    const alreadyInList = filenames.includes(file.filename)
-    const anyFile = file.comments.filter(({ structure }) => structure.asFile)
-    const someFile = anyFile.length > 0 ? anyFile[0] : false
-    const asFilePath = pipe(
-      defaultTo({}),
-      pathOr('???', ['structure', 'asFile'])
-    )(someFile)
-    const withOrder = pipe(pathOr('0', ['structure', 'order']), x =>
-      parseInt(x, 10)
-    )(someFile)
-    const dir = dirname(file.filename)
-    const newFile = someFile ? pathJoin(dir, asFilePath) : '???'
-    return alreadyInList
-      ? map(raw => {
-          const check = raw.filename === file.filename
-          return check ? combineFiles(raw.order < withOrder, raw, file) : raw
-        })(agg)
-      : [
-          ...agg,
-          someFile
-            ? {
-                ...file,
-                filename: newFile,
-                order: withOrder,
-                originalFilename: file.filename,
-              }
-            : file,
-        ]
-  }, [])
-)
-
-export const pullPageTitleFromAnyComment = pipe(
-  filter(pathOr(false, ['structure', 'page'])),
-  map(path(['structure', 'page'])),
-  head,
-  defaultTo(''),
-  replace(/\s/g, '-'),
-  defaultTo(false)
-)
-
-export const cleanFilename = curry(
-  (testMode, { fileGroup, filename, comments }) => {
-    const title = pullPageTitleFromAnyComment(comments)
-    const sliced = title || slug(filename)
-    const result = toLower(capitalToKebab(sliced)) + '.mdx'
-    return testMode
-      ? ''
-      : stripLeadingHyphen((fileGroup ? fileGroup + '/' : '') + result)
-  }
+export const processComments = curry((testMode, x) =>
+  (testMode ? filterAndStructureTests : filterAndStructureComments)(x)
 )
