+++ packages/doctor-general/doctor-general.js
@@ -1,9 +1,10 @@
 export * from './comment'
+export * from './constants'
 export * from './drgen'
 export * from './file'
+export * from './log'
 export * from './parse'
-export * from './renderer-jest'
-export * from './renderer-markdown'
+export * from './reader'
 export * from './stats'
-export * from './string'
 export * from './text'
+export * from './writer'
