--- packages/doctor-general/trace.js
@@ -1,11 +1,0 @@
-import { curry, __ as $, identity as I } from 'ramda'
-
-export const scopedBinaryEffect = curry((effect, scope, a, b) => {
-  effect(a, scope(b))
-  return b
-})
-
-export const binaryEffect = scopedBinaryEffect($, I)
-
-// eslint-disable-next-line no-console
-export const trace = binaryEffect(console.log)
